requirements.txt
PySide6
pyinstaller
Pillow
cairosvg